python_service/requirements.txt
fastapi
grpcio==1.62.1
grpcio-tools==1.62.1
protobuf==4.25.3
psutil
uvicorn
